Extract common types in upload service to upload types Required for not including code required for upload service in other parts just requiring common types e.g. ffmpegService is included in machine learning web worker while including downloadManager service
