+++ src/services/upload/encryptionService.ts
@@ -1,4 +1,5 @@
-import { DataStream, EncryptionResult, isDataStream } from './uploadService';
+import { DataStream, EncryptionResult } from 'types/upload';
+import { isDataStream } from './uploadService';
 
 async function encryptFileStream(worker, fileData: DataStream) {
     const { stream, chunkCount } = fileData;
